drivers: pci: pcie_brcmstb: set correct reset state on pcie_remove
For the controllers that are utilizing reset for bridge sw init,
such as BCM2712, reset should be deasserted after assertion.
Leaving it in asserted state may lead to unexpected hangs in
the Linux Kernel driver because it do not perform reset initialization
and start accessing the device memory.

We've met this issue on the Raspberry PI5 kernel and got the following
log:

[    1.842953] brcm-pcie 1000110000.pcie: host bridge /axi/pcie@110000 ranges:
[    1.849953] brcm-pcie 1000110000.pcie:   No bus range found for /axi/pcie@110000, using [bus 00-ff]
[    1.859104] brcm-pcie 1000110000.pcie:      MEM 0x1b00000000..0x1bfffffffb -> 0x0000000000
[    1.867407] brcm-pcie 1000110000.pcie:      MEM 0x1800000000..0x1affffffff -> 0x0400000000
[    1.875707] brcm-pcie 1000110000.pcie:   IB MEM 0x0000000000..0x0fffffffff -> 0x1000000000nc

After that system reports a context switch error after 120 sec timeout.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Tested-by: Etienne Dublé <etienne.duble@imag.fr>