bcm2712: enable linux kernel image header
This change is needed to make OP-TEE operable on RPI5.

This is somewhat convoluted, but in the essence, OP-TEE image is being
loaded at address 0x80000 and we can't change that. Problem is that
last versions (like 2024/06/05) of RPI5 firmware tries to load
non-kernel images to the same address 0x80000, thus, overwriting
OP-TEE image.

This patch enables Linux kernel image for U-Boot, so firmware will
treat U-Boot as a kernel image, placing it at different address,
namely at 0x200000.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Tested-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Tested-by: Etienne Dublé <etienne.duble@imag.fr>